Connection to rest of program
I've connected the fidget toys to the rest of the program. I have done this by add routes for all the different fidget toy. have edit the fidget spinner so looks cleaner.
Edit some of the css for the spacing to not to big ,eg padding and margin for the form element